requirements.txt
streamlit
requests
beautifulsoup4
openai
pillow
python-dotenv
tweepy
instagrapi 